Gallery

+++ app/Http/Controllers/UsersController.php
@@ -13,7 +13,9 @@ use Illuminate\Support\Facades\Storage;
 
 class UsersController extends Controller
 {
-
+    /*
+     * Avatar files that cannot be deleted.
+     */
     private $avatars = [
                 '1' => 'avatar/avatar1.png',
                 '2' => 'avatar/avatar2.png',
@@ -29,19 +31,12 @@ class UsersController extends Controller
             ];
 
     /**
-     * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        $user = User::where('id', Auth::id())->first();
-        $usersIds = $user->following()->pluck('follow_id')->all();
-        //$usersIds[] = $user->id;
-        $pictures = Picture::whereIn('user_id', $usersIds)->where('accept', 1)->latest()->paginate(10);
-
-
-        return view('users.index', compact('pictures'));
+        //
     }
 
     /**
@@ -125,7 +120,7 @@ class UsersController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update the user avatar from hard drive.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -133,47 +128,53 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if ($request->hasFile('avatar')) {
+        if (Auth::id() == $id) {
+            if ($request->hasFile('avatar')) {
 
-            $request->validate([
-                'avatar' => 'mimes:jpeg,bmp,png,jpg'
-            ]);
-            $path = $request->file('avatar')->store('avatar', 'public');
+                $request->validate([
+                    'avatar' => 'mimes:jpeg,bmp,png,jpg'
+                ]);
+                $path = $request->file('avatar')->store('avatar', 'public');
 
-            $user = User::find($id);
-            $oldfilename = $user->avatar;
+                $user = User::find($id);
+                $oldfilename = $user->avatar;
 
-            if (! in_array($oldfilename, $this->avatars) ){
-                Storage::disk('public')->delete($user->avatar);
+                if (! in_array($oldfilename, $this->avatars)) {
+                    Storage::disk('public')->delete($user->avatar);
+                }
+
+                $user->avatar = $path;
+                $user->save();
+
+
+                return redirect()->back();
             }
-
-            $user->avatar = $path;
-            $user->save();
-
+        } else {
             return redirect()->back();
         }
     }
 
     /**
-     * Update the specified resource in storage.
+     * User set avatar from default avatars.
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function defaultAvatar($id, $x)
     {
+        if (Auth::id() == $id) {
+            $user = User::find($id);
+            $oldfilename = $user->avatar;
 
-        $user = User::find($id);
-        $oldfilename = $user->avatar;
+            if (!in_array($oldfilename, $this->avatars)) {
+                Storage::disk('public')->delete($user->avatar);
+            }
+            $user->avatar = 'avatar/avatar' . $x . '.png';
+            $user->save();
 
-
-        if (! in_array($oldfilename, $this->avatars) ){
-            Storage::disk('public')->delete($user->avatar);
+            return redirect()->back();
+        } else {
+            return redirect()->back();
         }
-        $user->avatar = 'avatar/avatar' . $x . '.png';
-
-        $user->save();
-
-        return redirect()->back();
 
     }
